Refactor visualization and eval code
